Reduce Media/File picker samples
